refactor: convert revenue controller to Dashboard
