+++ src/app.module.ts
@@ -1,47 +1,47 @@
 import { MailerModule } from '@nestjs-modules/mailer';
 import { HandlebarsAdapter } from '@nestjs-modules/mailer/dist/adapters/handlebars.adapter';
+import { BullModule } from '@nestjs/bull';
 import { Module, ValidationPipe } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { APP_PIPE } from '@nestjs/core';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { BullModule } from '@nestjs/bull';
 import * as Joi from 'joi';
 import { LoggerModule } from 'nestjs-pino';
+import authConfig from '../config/auth.config';
 import serverConfig from '../config/server.config';
 import dataSource from './database/data-source';
 import { SeedingModule } from './database/seeding/seeding.module';
+import { AuthGuard } from './guards/auth.guard';
 import HealthController from './health.controller';
 import { AuthModule } from './modules/auth/auth.module';
+import { BillingPlanModule } from './modules/billing-plans/billing-plan.module';
+import { BlogModule } from './modules/blogs/blogs.module';
+import { ContactUsModule } from './modules/contact-us/contact-us.module';
+import { RevenueModule } from './modules/dashboard/dashboard.module';
 import { EmailModule } from './modules/email/email.module';
-import { JobsModule } from './modules/jobs/jobs.module';
+import { FaqModule } from './modules/faq/faq.module';
+import { FlutterwaveModule } from './modules/flutterwave/flutterwave.module';
+import { HelpCenterModule } from './modules/help-center/help-center.module';
 import { InviteModule } from './modules/invite/invite.module';
+import { JobsModule } from './modules/jobs/jobs.module';
+import { NewsletterSubscriptionModule } from './modules/newsletter-subscription/newsletter-subscription.module';
+import { NotificationSettingsModule } from './modules/notification-settings/notification-settings.module';
+import { NotificationsModule } from './modules/notifications/notifications.module';
+import { OrganisationPermissionsModule } from './modules/organisation-permissions/organisation-permissions.module';
+import { OrganisationRoleModule } from './modules/organisation-role/organisation-role.module';
 import { OrganisationsModule } from './modules/organisations/organisations.module';
 import { OtpModule } from './modules/otp/otp.module';
-import authConfig from '../config/auth.config';
-import { AuthGuard } from './guards/auth.guard';
 import { ProductsModule } from './modules/products/products.module';
-import { BillingPlanModule } from './modules/billing-plans/billing-plan.module';
-import { NotificationSettingsModule } from './modules/notification-settings/notification-settings.module';
 import { ProfileModule } from './modules/profile/profile.module';
 import { SqueezeModule } from './modules/squeeze/squeeze.module';
+import { SubscriptionsModule } from './modules/subscriptions/subscriptions.module';
+import { TeamsModule } from './modules/teams/teams.module';
 import { TestimonialsModule } from './modules/testimonials/testimonials.module';
 import { TimezonesModule } from './modules/timezones/timezones.module';
 import { UserModule } from './modules/user/user.module';
+import { WaitlistModule } from './modules/waitlist/waitlist.module';
 import ProbeController from './probe.controller';
 import { RunTestsModule } from './run-tests/run-tests.module';
-import { NotificationsModule } from './modules/notifications/notifications.module';
-import { OrganisationRoleModule } from './modules/organisation-role/organisation-role.module';
-import { OrganisationPermissionsModule } from './modules/organisation-permissions/organisation-permissions.module';
-import { ContactUsModule } from './modules/contact-us/contact-us.module';
-import { FaqModule } from './modules/faq/faq.module';
-import { HelpCenterModule } from './modules/help-center/help-center.module';
-import { WaitlistModule } from './modules/waitlist/waitlist.module';
-import { NewsletterSubscriptionModule } from './modules/newsletter-subscription/newsletter-subscription.module';
-import { TeamsModule } from './modules/teams/teams.module';
-import { FlutterwaveModule } from './modules/flutterwave/flutterwave.module';
-import { BlogModule } from './modules/blogs/blogs.module';
-import { SubscriptionsModule } from './modules/subscriptions/subscriptions.module';
-import { RevenueModule } from './modules/revenue/revenue.module';
 
 @Module({
   providers: [
