+++ src/modules/dashboard/dashboard.module.ts
@@ -1,15 +1,15 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { DashboardController } from './dashboard.controller';
+import { DashboardService } from './dashboard.service';
 import { Cart } from './entities/cart.entity';
 import { OrderItem } from './entities/order-items.entity';
 import { Order } from './entities/order.entity';
 import { Transaction } from './entities/transaction.entity';
-import { RevenueController } from './revenue.controller';
-import { RevenueService } from './revenue.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Transaction, Order, OrderItem, Cart])],
-  controllers: [RevenueController],
-  providers: [RevenueService],
+  controllers: [DashboardController],
+  providers: [DashboardService],
 })
 export class RevenueModule {}
