+++ src/modules/dashboard/entities/cart.entity.ts
